refactor: consolidate auto-compact functionality into autocompact tool
- Move all functions from gptme/util/auto_compact.py to gptme/tools/autocompact.py
- Move /compact command from commands.py to autocompact tool
- Register /compact command via tool's commands dict
- Remove duplicate code from commands.py
- Delete gptme/util/auto_compact.py
- Update test imports to new location
- Fix logger.exception -> logger.error in hooks.py

This eliminates duplication and consolidates all auto-compacting
functionality in a single location, following the pattern of other
tools that register commands.
